Create blurTest3.py
Latest attempt at the MATLAB port. I think it is good up until the graph theory part... but it is awfully hard to test.  The blur algorithms are (as of the this version) very similar, but not quite 100% doing the same thing.

Next up: figure out how to make dijkstra.py take in my sparse vector representation and/or come up with a between sparse array system (233121 x 233121 with only 233121 non-zero elements)